# Conflicts:
#	testcode/test_pool.py
#	workplace/forone/get_standard_data.py
#	workplace/forone/relation_category_keywords.py

+++ workplace/forone/count_category_num.py
@@ -10,7 +10,7 @@ def count_the_number_of_categories(csv_data):
     new_word_list = list(set(word_list))
     # new_word_list.sort(key=word_list.index)
     print(new_word_list[:10])
-    dummies = pd.DataFrame(np.zeros((len(csv_data), len(new_word_list))), columns=new_word_list)
+    dummies = pd.DataFrame(np.zeros((len(csv_data), len(new_word_list)), dtype=np.int8), columns=new_word_list)
     for index in range(0, len(csv_data)):
         for word in csv_data['cut_name'].iloc[index]:
             if word in new_word_list:
